+++ src/routes/settings/+page.server.ts
@@ -2,6 +2,6 @@ import { redirect } from "@sveltejs/kit";
 import type { PageServerLoad } from "./$types";
 
 export const load: PageServerLoad = async () => {
-	// Redirect to profile as the default settings page
-	redirect(302, "/settings/profile");
+	// Redirect to general as the default settings page
+	redirect(302, "/settings/general");
 };